refactor: 버거 목록 페이지 컴포넌트 분리
