+++ src/components/BurgerList/BurgerBanner.tsx
@@ -1,0 +1,13 @@
+import * as S from './BurgerBanner.style';
+import { Cimg12 } from '@assets/svgs/burgerList';
+
+const BurgerBanner = () => {
+  return (
+    <section css={S.Banner}>
+      <h1>버거</h1>
+      <Cimg12 />
+    </section>
+  );
+};
+
+export default BurgerBanner;
